requirements.txt
aiofile==3.9.0 ; python_version >= "3.10" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.11.0 ; python_version >= "3.10" and python_version < "4.0"
caio==0.9.24 ; python_version >= "3.10" and python_version < "4.0"
certifi==2025.10.5 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.4.3 ; python_version >= "3.10" and python_version < "4.0"
click==8.3.0 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
exceptiongroup==1.3.0 ; python_version == "3.10"
fastapi-cache2==0.2.2 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.115.14 ; python_version >= "3.10" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==6.11.0 ; python_version >= "3.10" and python_version < "4.0"
mangum==0.17.0 ; python_version >= "3.10" and python_version < "4.0"
markdown==3.9 ; python_version >= "3.10" and python_version < "4.0"
numpy==2.2.6 ; python_version == "3.10"
numpy==2.3.3 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.3.3 ; python_version >= "3.10" and python_version < "4.0"
pendulum==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.41.1 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.12.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.3 ; python_version >= "3.10" and python_version < "4.0"
rapidfuzz==3.14.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.10" and python_version < "4.0"
scipy==1.15.3 ; python_version == "3.10"
scipy==1.16.2 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.46.2 ; python_version >= "3.10" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.10" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
xmltodict==1.0.2 ; python_version >= "3.10" and python_version < "4.0"
zipp==3.23.0 ; python_version >= "3.10" and python_version < "4.0"
